MSX Methods changed to work with 'int' and not 'unsigned 8-bit int' anymore. Sample programs changed to pass values using 'int' and not 'byte'. Created a new sample program to test VDP access. Yey!

+++ java/net/mikekohn/java_grinder/MSX.java
@@ -20,29 +20,29 @@ abstract public class MSX
 {
   protected MSX() { }
 
-  public static void backColor(byte color) { }
+  public static void backColor(int color) { }
   public static void beep() { }
-  public static void borderColor(byte color) { }
+  public static void borderColor(int color) { }
   public static void cls() { }
-  public static void color(byte foreground, byte background, byte border) { }
+  public static void color(int foreground, int background, int border) { }
   public static void copyVRAM(int source, int dest, int size) { }
   public static void fillVRAM(int value, int addr, int size) { }
-  public static void foreColor(byte color) { }
+  public static void foreColor(int color) { }
   public static int  getChar() { return 0; }
-  public static int  getStick(byte joy) { return 0; }
-  public static int  getTrig(byte joy) { return 0; }
+  public static int  getStick(int joy) { return 0; }
+  public static int  getTrig(int joy) { return 0; }
   public static void keyOff() { }
   public static void keyOn() { }
   public static void putChar(char c) { }
   // not working
   public static void putS(int i, char[] s) { }
   // not working
-  public static int  readVDP(byte reg) { return 0; }
+  public static int  readVDP(int reg) { return 0; }
   public static int  readVRAM(int addr) { return 0; }
-  public static void screen(byte mode) { }
-  public static void setCursor(byte column, byte line) { }
-  public static void width(byte w) { }
-  public static void writePSG(byte reg, byte value) { }
-  public static void writeVDP(byte reg, byte value) { }
-  public static void writeVRAM(int addr, byte value) { }
+  public static void screen(int mode) { }
+  public static void setCursor(int column, int line) { }
+  public static void width(int w) { }
+  public static void writePSG(int reg, int value) { }
+  public static void writeVDP(int reg, int value) { }
+  public static void writeVRAM(int addr, int value) { }
 }
